clean up, add to remaining pages
