+++ apps/web/src/components/AdminContentWrapper/AdminContentWrapper.tsx
@@ -1,24 +1,13 @@
 import React from "react";
 
-import { Breadcrumbs, BreadcrumbsProps } from "@gc-digital-talent/ui";
-
 interface AdminContentWrapperProps {
-  crumbs?: BreadcrumbsProps["crumbs"];
   children: React.ReactNode;
 }
 
-const AdminContentWrapper = ({
-  crumbs,
-  children,
-}: AdminContentWrapperProps) => {
-  return (
-    <>
-      <div data-h2-container="base(center, full, x2)">
-        <div data-h2-padding="base(0, 0, x3, 0)">{children}</div>
-      </div>
-      {crumbs && <Breadcrumbs crumbs={crumbs} />}
-    </>
-  );
-};
+const AdminContentWrapper = ({ children }: AdminContentWrapperProps) => (
+  <div data-h2-container="base(center, full, x2)">
+    <div data-h2-padding="base(0, 0, x3, 0)">{children}</div>
+  </div>
+);
 
 export default AdminContentWrapper;
