+++ apps/web/src/pages/Departments/CreateDepartmentPage.tsx
@@ -6,7 +6,6 @@ import { useIntl } from "react-intl";
 import { toast } from "@gc-digital-talent/toast";
 import { Input, Submit } from "@gc-digital-talent/forms";
 import { errorMessages } from "@gc-digital-talent/i18n";
-import { Heading } from "@gc-digital-talent/ui";
 
 import SEO from "~/components/SEO/SEO";
 import useRoutes from "~/hooks/useRoutes";
@@ -17,6 +16,7 @@ import {
 } from "~/api/generated";
 import AdminContentWrapper from "~/components/AdminContentWrapper/AdminContentWrapper";
 import adminMessages from "~/messages/adminMessages";
+import AdminHero from "~/components/Hero/AdminHero";
 
 type FormValues = CreateDepartmentInput;
 
@@ -68,13 +68,6 @@ export const CreateDepartmentForm = ({
 
   return (
     <section data-h2-container="base(left, s)">
-      <Heading level="h1" size="h2">
-        {intl.formatMessage({
-          defaultMessage: "Create Department",
-          id: "XBY4Fq",
-          description: "Title displayed on the create a department form.",
-        })}
-      </Heading>
       <FormProvider {...methods}>
         <form
           onSubmit={handleSubmit(onSubmit)}
@@ -169,17 +162,23 @@ const CreateDepartmentPage = () => {
     },
   ];
 
+  const pageTitle = intl.formatMessage({
+    defaultMessage: "Create department",
+    id: "PRrRRN",
+    description: "Page title for the department creation page",
+  });
+
   return (
-    <AdminContentWrapper crumbs={navigationCrumbs}>
-      <SEO
-        title={intl.formatMessage({
-          defaultMessage: "Create department",
-          id: "PRrRRN",
-          description: "Page title for the department creation page",
-        })}
+    <>
+      <SEO title={pageTitle} />
+      <AdminHero
+        title={pageTitle}
+        nav={{ mode: "crumbs", items: navigationCrumbs }}
       />
-      <CreateDepartmentForm handleCreateDepartment={handleCreateDepartment} />
-    </AdminContentWrapper>
+      <AdminContentWrapper>
+        <CreateDepartmentForm handleCreateDepartment={handleCreateDepartment} />
+      </AdminContentWrapper>
+    </>
   );
 };
