+++ apps/web/src/pages/Pools/ViewPoolPage/ViewPoolPage.tsx
@@ -4,7 +4,7 @@ import { useIntl } from "react-intl";
 import UserGroupIcon from "@heroicons/react/24/outline/UserGroupIcon";
 
 import { Pending, NotFound, Link, Heading, Pill } from "@gc-digital-talent/ui";
-import { commonMessages, getLocalizedName } from "@gc-digital-talent/i18n";
+import { commonMessages } from "@gc-digital-talent/i18n";
 import {
   DATE_FORMAT_STRING,
   formatDate,
@@ -23,7 +23,6 @@ import SEO from "~/components/SEO/SEO";
 import useRoutes from "~/hooks/useRoutes";
 import useRequiredParams from "~/hooks/useRequiredParams";
 import AdminContentWrapper from "~/components/AdminContentWrapper/AdminContentWrapper";
-import adminMessages from "~/messages/adminMessages";
 import ProcessCard from "~/components/ProcessCard/ProcessCard";
 import {
   getAdvertisementStatus,
@@ -414,38 +413,14 @@ type RouteParams = {
 
 const ViewPoolPage = () => {
   const intl = useIntl();
-  const routes = useRoutes();
   const { poolId } = useRequiredParams<RouteParams>("poolId");
   const { isFetching, mutations } = usePoolMutations();
   const [{ data, fetching, error }] = useGetProcessInfoQuery({
     variables: { id: poolId },
   });
 
-  const navigationCrumbs = [
-    {
-      label: intl.formatMessage({
-        defaultMessage: "Home",
-        id: "EBmWyo",
-        description: "Link text for the home link in breadcrumbs.",
-      }),
-      url: routes.home(),
-    },
-    {
-      label: intl.formatMessage(adminMessages.pools),
-      url: routes.poolTable(),
-    },
-    ...(poolId
-      ? [
-          {
-            label: getLocalizedName(data?.pool?.name, intl),
-            url: routes.poolView(poolId),
-          },
-        ]
-      : []),
-  ];
-
   return (
-    <AdminContentWrapper crumbs={navigationCrumbs}>
+    <AdminContentWrapper>
       <Pending fetching={fetching} error={error}>
         {poolId && data?.pool ? (
           <ViewPool
