+++ apps/web/src/pages/SearchRequests/ViewSearchRequestPage/ViewSearchRequestPage.tsx
@@ -1,10 +1,7 @@
 import React from "react";
-import { useIntl } from "react-intl";
 
-import SEO from "~/components/SEO/SEO";
 import { Scalars } from "~/api/generated";
 import useRequiredParams from "~/hooks/useRequiredParams";
-import AdminHero from "~/components/Hero/AdminHero";
 
 import ViewSearchRequestApi from "./components/ViewSearchRequest";
 
@@ -13,21 +10,9 @@ type RouteParams = {
 };
 
 export const SingleSearchRequestPage = () => {
-  const intl = useIntl();
   const { searchRequestId } = useRequiredParams<RouteParams>("searchRequestId");
-  const pageTitle = intl.formatMessage({
-    defaultMessage: "Request",
-    id: "WYJnLs",
-    description: "Heading displayed above the single search request component.",
-  });
 
-  return (
-    <>
-      <SEO title={pageTitle} />
-      <AdminHero title={pageTitle} />
-      <ViewSearchRequestApi searchRequestId={searchRequestId} />
-    </>
-  );
+  return <ViewSearchRequestApi searchRequestId={searchRequestId} />;
 };
 
 export default SingleSearchRequestPage;
