+++ apps/web/src/pages/Skills/UpdateSkillPage.tsx
@@ -22,7 +22,7 @@ import {
   commonMessages,
   getSkillCategory,
 } from "@gc-digital-talent/i18n";
-import { Pending, NotFound, Heading } from "@gc-digital-talent/ui";
+import { Pending, NotFound } from "@gc-digital-talent/ui";
 
 import SEO from "~/components/SEO/SEO";
 import {
@@ -40,6 +40,7 @@ import useRequiredParams from "~/hooks/useRequiredParams";
 import AdminContentWrapper from "~/components/AdminContentWrapper/AdminContentWrapper";
 import adminMessages from "~/messages/adminMessages";
 import { parseKeywords } from "~/utils/skillUtils";
+import AdminHero from "~/components/Hero/AdminHero";
 
 type Option<V> = { value: V; label: string };
 
@@ -145,158 +146,148 @@ export const UpdateSkillForm = ({
 
   return (
     <section data-h2-container="base(left, s)">
-      <Heading level="h1" size="h2">
-        {intl.formatMessage({
-          defaultMessage: "Update Skill",
-          id: "WoAuKA",
-          description: "Title displayed on the update a skill form.",
-        })}
-      </Heading>
-      <div>
-        <FormProvider {...methods}>
-          <form
-            onSubmit={handleSubmit(onSubmit)}
-            data-h2-display="base(flex)"
-            data-h2-flex-direction="base(column)"
-            data-h2-gap="base(x1 0)"
-          >
-            <Input
-              id="name_en"
-              name="name.en"
-              label={intl.formatMessage({
-                defaultMessage: "Name (English)",
-                id: "5F/aKm",
-                description:
-                  "Label displayed on the update a skill form name (English) field.",
-              })}
-              type="text"
-              rules={{
-                required: intl.formatMessage(errorMessages.required),
-              }}
-            />
-            <Input
-              id="name_fr"
-              name="name.fr"
-              label={intl.formatMessage({
-                defaultMessage: "Name (French)",
-                id: "j01W+1",
-                description:
-                  "Label displayed on the update a skill form name (French) field.",
-              })}
-              type="text"
-              rules={{
-                required: intl.formatMessage(errorMessages.required),
-              }}
-            />
-            <TextArea
-              id="description_en"
-              name="description.en"
-              label={intl.formatMessage({
-                defaultMessage: "Description (English)",
-                id: "fdKtYm",
-                description:
-                  "Label displayed on the update a skill form description (English) field.",
-              })}
-              rules={{
-                required: intl.formatMessage(errorMessages.required),
-              }}
-            />
-            <TextArea
-              id="description_fr"
-              name="description.fr"
-              label={intl.formatMessage({
-                defaultMessage: "Description (French)",
-                id: "4EkI/1",
-                description:
-                  "Label displayed on the update a skill form description (French) field.",
-              })}
-              rules={{
-                required: intl.formatMessage(errorMessages.required),
-              }}
-            />
-            <Input
-              id="keywords_en"
-              name="keywords.en"
-              label={intl.formatMessage({
-                defaultMessage: "Keywords (English)",
-                id: "FiylOa",
-                description:
-                  "Label displayed on the skill form keywords field in English.",
-              })}
-              context={intl.formatMessage({
-                defaultMessage:
-                  "This field accepts a list of comma separated keywords associated with the skill.",
-                id: "NT3jrI",
-                description:
-                  "Additional context describing the purpose of the skills 'keyword' field.",
-              })}
-              type="text"
-            />
-            <Input
-              id="keywords_fr"
-              name="keywords.fr"
-              label={intl.formatMessage({
-                defaultMessage: "Keywords (French)",
-                id: "fOl4Ez",
-                description:
-                  "Label displayed on the skill form keywords field in French.",
-              })}
-              context={intl.formatMessage({
-                defaultMessage:
-                  "This field accepts a list of comma separated keywords associated with the skill.",
-                id: "NT3jrI",
-                description:
-                  "Additional context describing the purpose of the skills 'keyword' field.",
-              })}
-              type="text"
-            />
-            <Select
-              id="category"
-              name="category"
-              label={intl.formatMessage({
-                defaultMessage: "Category",
-                id: "KZR3ad",
-                description:
-                  "Label displayed on the skill family form category field.",
-              })}
-              nullSelection={intl.formatMessage({
-                defaultMessage: "Select a category",
-                id: "+hRCVl",
-                description:
-                  "Placeholder displayed on the skill family form category field.",
-              })}
-              rules={{
-                required: intl.formatMessage(errorMessages.required),
-              }}
-              options={enumToOptions(SkillCategory).map(({ value }) => ({
-                value,
-                label: intl.formatMessage(getSkillCategory(value)),
-              }))}
-            />
-            <Combobox
-              id="families"
-              name="families"
-              isMulti
-              label={intl.formatMessage({
-                defaultMessage: "Families",
-                id: "JxVREd",
-                description:
-                  "Label displayed on the skill form families field.",
-              })}
-              placeholder={intl.formatMessage({
-                defaultMessage: "Select one or more families",
-                id: "wORNl0",
-                description:
-                  "Placeholder displayed on the skill form families field.",
-              })}
-              options={skillFamilyOptions}
-            />
-            <div data-h2-align-self="base(flex-start)">
-              <Submit />
-            </div>
-          </form>
-        </FormProvider>
-      </div>
+      <FormProvider {...methods}>
+        <form
+          onSubmit={handleSubmit(onSubmit)}
+          data-h2-display="base(flex)"
+          data-h2-flex-direction="base(column)"
+          data-h2-gap="base(x1 0)"
+        >
+          <Input
+            id="name_en"
+            name="name.en"
+            label={intl.formatMessage({
+              defaultMessage: "Name (English)",
+              id: "5F/aKm",
+              description:
+                "Label displayed on the update a skill form name (English) field.",
+            })}
+            type="text"
+            rules={{
+              required: intl.formatMessage(errorMessages.required),
+            }}
+          />
+          <Input
+            id="name_fr"
+            name="name.fr"
+            label={intl.formatMessage({
+              defaultMessage: "Name (French)",
+              id: "j01W+1",
+              description:
+                "Label displayed on the update a skill form name (French) field.",
+            })}
+            type="text"
+            rules={{
+              required: intl.formatMessage(errorMessages.required),
+            }}
+          />
+          <TextArea
+            id="description_en"
+            name="description.en"
+            label={intl.formatMessage({
+              defaultMessage: "Description (English)",
+              id: "fdKtYm",
+              description:
+                "Label displayed on the update a skill form description (English) field.",
+            })}
+            rules={{
+              required: intl.formatMessage(errorMessages.required),
+            }}
+          />
+          <TextArea
+            id="description_fr"
+            name="description.fr"
+            label={intl.formatMessage({
+              defaultMessage: "Description (French)",
+              id: "4EkI/1",
+              description:
+                "Label displayed on the update a skill form description (French) field.",
+            })}
+            rules={{
+              required: intl.formatMessage(errorMessages.required),
+            }}
+          />
+          <Input
+            id="keywords_en"
+            name="keywords.en"
+            label={intl.formatMessage({
+              defaultMessage: "Keywords (English)",
+              id: "FiylOa",
+              description:
+                "Label displayed on the skill form keywords field in English.",
+            })}
+            context={intl.formatMessage({
+              defaultMessage:
+                "This field accepts a list of comma separated keywords associated with the skill.",
+              id: "NT3jrI",
+              description:
+                "Additional context describing the purpose of the skills 'keyword' field.",
+            })}
+            type="text"
+          />
+          <Input
+            id="keywords_fr"
+            name="keywords.fr"
+            label={intl.formatMessage({
+              defaultMessage: "Keywords (French)",
+              id: "fOl4Ez",
+              description:
+                "Label displayed on the skill form keywords field in French.",
+            })}
+            context={intl.formatMessage({
+              defaultMessage:
+                "This field accepts a list of comma separated keywords associated with the skill.",
+              id: "NT3jrI",
+              description:
+                "Additional context describing the purpose of the skills 'keyword' field.",
+            })}
+            type="text"
+          />
+          <Select
+            id="category"
+            name="category"
+            label={intl.formatMessage({
+              defaultMessage: "Category",
+              id: "KZR3ad",
+              description:
+                "Label displayed on the skill family form category field.",
+            })}
+            nullSelection={intl.formatMessage({
+              defaultMessage: "Select a category",
+              id: "+hRCVl",
+              description:
+                "Placeholder displayed on the skill family form category field.",
+            })}
+            rules={{
+              required: intl.formatMessage(errorMessages.required),
+            }}
+            options={enumToOptions(SkillCategory).map(({ value }) => ({
+              value,
+              label: intl.formatMessage(getSkillCategory(value)),
+            }))}
+          />
+          <Combobox
+            id="families"
+            name="families"
+            isMulti
+            label={intl.formatMessage({
+              defaultMessage: "Families",
+              id: "JxVREd",
+              description: "Label displayed on the skill form families field.",
+            })}
+            placeholder={intl.formatMessage({
+              defaultMessage: "Select one or more families",
+              id: "wORNl0",
+              description:
+                "Placeholder displayed on the skill form families field.",
+            })}
+            options={skillFamilyOptions}
+          />
+          <div data-h2-align-self="base(flex-start)">
+            <Submit />
+          </div>
+        </form>
+      </FormProvider>
     </section>
   );
 };
@@ -363,40 +354,46 @@ export const UpdateSkill = () => {
       : []),
   ];
 
+  const pageTitle = intl.formatMessage({
+    defaultMessage: "Edit skill",
+    id: "0VlOQq",
+    description: "Page title for the edit skill page.",
+  });
+
   return (
-    <AdminContentWrapper crumbs={navigationCrumbs}>
-      <SEO
-        title={intl.formatMessage({
-          defaultMessage: "Edit skill",
-          id: "0VlOQq",
-          description: "Page title for the edit skill page.",
-        })}
+    <>
+      <SEO title={pageTitle} />
+      <AdminHero
+        title={pageTitle}
+        nav={{ mode: "crumbs", items: navigationCrumbs }}
       />
-      <Pending fetching={fetching} error={error}>
-        {lookupData?.skill ? (
-          <UpdateSkillForm
-            initialSkill={lookupData?.skill}
-            families={families}
-            handleUpdateSkill={handleUpdateSkill}
-          />
-        ) : (
-          <NotFound
-            headingMessage={intl.formatMessage(commonMessages.notFound)}
-          >
-            <p>
-              {intl.formatMessage(
-                {
-                  defaultMessage: "Skill {skillId} not found.",
-                  id: "953EAy",
-                  description: "Message displayed for skill not found.",
-                },
-                { skillId },
-              )}
-            </p>
-          </NotFound>
-        )}
-      </Pending>
-    </AdminContentWrapper>
+      <AdminContentWrapper>
+        <Pending fetching={fetching} error={error}>
+          {lookupData?.skill ? (
+            <UpdateSkillForm
+              initialSkill={lookupData?.skill}
+              families={families}
+              handleUpdateSkill={handleUpdateSkill}
+            />
+          ) : (
+            <NotFound
+              headingMessage={intl.formatMessage(commonMessages.notFound)}
+            >
+              <p>
+                {intl.formatMessage(
+                  {
+                    defaultMessage: "Skill {skillId} not found.",
+                    id: "953EAy",
+                    description: "Message displayed for skill not found.",
+                  },
+                  { skillId },
+                )}
+              </p>
+            </NotFound>
+          )}
+        </Pending>
+      </AdminContentWrapper>
+    </>
   );
 };
