+++ apps/web/src/pages/Users/UserInformationPage/UserInformationPage.tsx
@@ -11,9 +11,7 @@ import { notEmpty } from "@gc-digital-talent/helpers";
 import SEO from "~/components/SEO/SEO";
 import { Scalars, useGetViewUserDataQuery } from "~/api/generated";
 import AdminContentWrapper from "~/components/AdminContentWrapper/AdminContentWrapper";
-import useRoutes from "~/hooks/useRoutes";
 import useRequiredParams from "~/hooks/useRequiredParams";
-import { getFullNameLabel } from "~/utils/nameUtils";
 import adminMessages from "~/messages/adminMessages";
 
 import AboutSection from "./components/AboutSection";
@@ -110,7 +108,6 @@ type RouteParams = {
 const UserInformationPage = () => {
   const { userId } = useRequiredParams<RouteParams>("userId");
   const intl = useIntl();
-  const routes = useRoutes();
   const [{ data: lookupData, fetching, error }] = useGetViewUserDataQuery({
     variables: { id: userId },
   });
@@ -118,31 +115,8 @@ const UserInformationPage = () => {
   const user = lookupData?.user;
   const pools = lookupData?.pools.filter(notEmpty);
 
-  const navigationCrumbs = [
-    {
-      label: intl.formatMessage({
-        defaultMessage: "Home",
-        id: "EBmWyo",
-        description: "Link text for the home link in breadcrumbs.",
-      }),
-      url: routes.adminDashboard(),
-    },
-    {
-      label: intl.formatMessage(adminMessages.users),
-      url: routes.userTable(),
-    },
-    ...(userId
-      ? [
-          {
-            label: getFullNameLabel(user?.firstName, user?.lastName, intl),
-            url: routes.userView(userId),
-          },
-        ]
-      : []),
-  ];
-
   return (
-    <AdminContentWrapper crumbs={navigationCrumbs}>
+    <AdminContentWrapper>
       <SEO
         title={intl.formatMessage({
           defaultMessage: "Candidate details",
